backend/requirements.txt
# Deep Learning & Data Science
tensorflow>=2.15.0
tensorflow-datasets>=4.9.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
seaborn>=0.12.0

# API & Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Image Processing
pillow>=10.0.0
opencv-python>=4.8.0

# Utilities
pydantic>=2.0.0
